+++ tests/testthat/test-crew_launcher.R
@@ -38,7 +38,7 @@ crew_test("launcher settings", {
   socket <- "ws://127.0.0.1:5000"
   settings <- launcher$settings(socket = socket)
   expect_equal(settings$url, socket)
-  expect_equal(settings$asyncdial, FALSE)
+  expect_equal(settings$autoexit, TRUE)
   expect_equal(settings$maxtasks, 7)
   expect_equal(settings$idletime, 2000)
   expect_equal(settings$walltime, 3000)